eventos-admin/requirements.txt
customtkinter
requests
qrcode
Pillow
pyshorteners 
python-dotenv
sqlalchemy
apscheduler
